+++ src/buffer/gaussian.rs
@@ -3,7 +3,7 @@ use glam::*;
 use half::f16;
 use wgpu::util::DeviceExt;
 
-use crate::{CameraTrait, Gaussian, wgpu_sort};
+use crate::{CameraTrait, Error, Gaussian, wgpu_sort};
 
 /// The Gaussians storage buffer.
 #[derive(Debug)]
@@ -483,7 +483,7 @@ pub struct GaussianShDegree(u8);
 impl GaussianShDegree {
     /// Create a new Gaussian SH degree.
     ///
-    /// Returns `None` if the degree is not in the range of \[0, 3\].
+    /// Returns [`None`] if the degree is not in the range of \[0, 3\].
     pub const fn new(sh_deg: u8) -> Option<Self> {
         match sh_deg {
             0..=3 => Some(Self(sh_deg)),
@@ -550,7 +550,7 @@ impl GaussianTransformBuffer {
 pub struct GaussianTransformPod {
     pub size: f32,
 
-    /// (display_mode, sh_deg, no_sh0, padding)
+    /// \[display_mode, sh_deg, no_sh0, padding\]
     pub flags: U8Vec4,
 }
 
@@ -612,26 +612,79 @@ impl GaussiansDepthBuffer {
 
 /// The Gaussians edit storage buffer.
 #[derive(Debug, Clone)]
-pub struct GaussiansEditBuffer(wgpu::Buffer);
+pub struct GaussiansEditBuffer {
+    data: wgpu::Buffer,
+    download: wgpu::Buffer,
+}
 
 impl GaussiansEditBuffer {
     /// Create a new Gaussians edit buffer.
     pub fn new(device: &wgpu::Device, gaussian_count: u32) -> Self {
         let size = gaussian_count * std::mem::size_of::<GaussianEditPod>() as u32;
 
-        let buffer = device.create_buffer(&wgpu::BufferDescriptor {
+        let data = device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("Gaussians Edit Buffer"),
             size: size as wgpu::BufferAddress,
-            usage: wgpu::BufferUsages::STORAGE,
+            usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_SRC,
             mapped_at_creation: false,
         });
 
-        Self(buffer)
+        let download = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("Gaussians Edit Download Buffer"),
+            size: size as wgpu::BufferAddress,
+            usage: wgpu::BufferUsages::COPY_DST | wgpu::BufferUsages::MAP_READ,
+            mapped_at_creation: false,
+        });
+
+        Self { data, download }
+    }
+
+    /// Download the Gaussian edit.
+    pub async fn download(
+        &self,
+        device: &wgpu::Device,
+        queue: &wgpu::Queue,
+    ) -> Result<Vec<GaussianEditPod>, Error> {
+        let mut encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
+            label: Some("Gaussian Edit Download Encoder"),
+        });
+        self.prepare_download(&mut encoder);
+        queue.submit(Some(encoder.finish()));
+
+        self.map_download(device).await
+    }
+
+    /// Prepare for downloading the Gaussian edit.
+    pub fn prepare_download(&self, encoder: &mut wgpu::CommandEncoder) {
+        encoder.copy_buffer_to_buffer(self.buffer(), 0, &self.download, 0, self.download.size());
+    }
+
+    /// Map the download buffer to read the Gaussian edit.
+    pub async fn map_download(&self, device: &wgpu::Device) -> Result<Vec<GaussianEditPod>, Error> {
+        let (tx, rx) = oneshot::channel();
+        let buffer_slice = self.download.slice(..);
+        buffer_slice.map_async(wgpu::MapMode::Read, move |result| {
+            if let Err(e) = tx.send(result) {
+                log::error!("Error occurred while sending Gaussian edit: {e:?}");
+            }
+        });
+        device.poll(wgpu::Maintain::Wait);
+        rx.await??;
+
+        let edits = bytemuck::allocation::pod_collect_to_vec(&buffer_slice.get_mapped_range());
+        self.download.unmap();
+
+        Ok(edits)
     }
 
     /// Get the buffer.
     pub fn buffer(&self) -> &wgpu::Buffer {
-        &self.0
+        &self.data
+    }
+
+    /// Get the download buffer.
+    pub fn download_buffer(&self) -> &wgpu::Buffer {
+        &self.download
     }
 }
 
@@ -701,6 +754,51 @@ impl GaussianEditPod {
             flag_hsv: u8vec4(flag, hsv_or_rgb.x, hsv_or_rgb.y, hsv_or_rgb.z),
             contr_expo_gamma_alpha,
         }
+    }
+
+    /// Get the flag.
+    pub fn flag(&self) -> GaussianEditFlag {
+        GaussianEditFlag::from_bits_truncate(self.flag_hsv.x)
+    }
+
+    /// Get the hue in the range of \[0, 1\].
+    pub fn hue(&self) -> f32 {
+        self.flag_hsv.y as f32 / 255.0
+    }
+
+    /// Get the saturation in the range of \[0, 2\].
+    pub fn saturation(&self) -> f32 {
+        self.flag_hsv.z as f32 / 127.5
+    }
+
+    /// Get the brightness in the range of \[0, 2\].
+    pub fn brightness(&self) -> f32 {
+        self.flag_hsv.w as f32 / 127.5
+    }
+
+    /// Get the RGB in the range of \[0, 1\].
+    pub fn rgb(&self) -> Vec3 {
+        self.flag_hsv.yzw().as_vec3().map(|x| x / 255.0)
+    }
+
+    /// Get the contrast in the range of \[-1, 1\].
+    pub fn contrast(&self) -> f32 {
+        self.contr_expo_gamma_alpha.x as f32 / 127.5 - 1.0
+    }
+
+    /// Get the exposure in the range of \[-5, 5\].
+    pub fn exposure(&self) -> f32 {
+        self.contr_expo_gamma_alpha.y as f32 / 25.5 - 5.0
+    }
+
+    /// Get the gamma in the range of \[0, 5\].
+    pub fn gamma(&self) -> f32 {
+        self.contr_expo_gamma_alpha.z as f32 / 51.0
+    }
+
+    /// Get the alpha in the range of \[0, 2\].
+    pub fn alpha(&self) -> f32 {
+        self.contr_expo_gamma_alpha.w as f32 / 127.5
     }
 }